refactor(configure): extract common helpers into ConfigureHelperTrait
Create ConfigureHelperTrait to eliminate code duplication across all
configure subcommands. Extracted three common methods used by multiple
subcommands into a reusable trait.

Extracted Methods:
- promptBoolean() - Interactive yes/no prompts with defaults
- parseBoolean()  - Parse string to boolean (true/false/yes/no/1/0/on/off)
- formatBoolean() - Format boolean for display (yes/no or custom text)

Applied to Subcommands:
- SessionHandler
- LDAP
- Groups
- Permissions
- Tokens

Code Reduction:
- Removed: 162 lines (duplicate implementations)
- Added: 69 lines (trait with documentation)
- Net: -93 lines of code

Benefits:
- Single source of truth for boolean operations
- Consistent behavior across all subcommands
- Easier maintenance (update once, applies everywhere)
- Enhanced formatBoolean() with customizable true/false text
- Better documentation in centralized location

All tests passing: 53 tests, 136 assertions

The trait follows the established pattern of using traits for shared
functionality (similar to HordectlModuleTrait).
