+++ src/Command/Configure/Ldap.php
@@ -39,6 +39,7 @@ use RuntimeException;
 class Ldap implements Module, ModuleUsage
 {
     use ModuleTrait;
+    use ConfigureHelperTrait;
 
     protected \Horde_Cli $cli;
     private ConfigManager $configManager;
@@ -599,42 +600,5 @@ class Ldap implements Module, ModuleUsage
                 $this->cli->writeln("    DN: {$entries[0]['dn']}");
             }
         }
-    }
-
-    /**
-     * Prompt for boolean value
-     *
-     * @param string $prompt Prompt text
-     * @param bool $default Default value
-     * @return bool User response
-     */
-    private function promptBoolean(string $prompt, bool $default): bool
-    {
-        $defaultStr = $default ? 'Y/n' : 'y/N';
-        $response = $this->cli->prompt("{$prompt} [{$defaultStr}]:", $default ? 'y' : 'n');
-        return strtolower($response) === 'y';
-    }
-
-    /**
-     * Parse boolean from string
-     *
-     * @param string $value String value
-     * @return bool Boolean value
-     */
-    private function parseBoolean(string $value): bool
-    {
-        $lower = strtolower($value);
-        return in_array($lower, ['true', '1', 'yes', 'y', 'on']);
-    }
-
-    /**
-     * Format boolean for display
-     *
-     * @param bool $value Boolean value
-     * @return string Formatted string
-     */
-    private function formatBoolean(bool $value): string
-    {
-        return $value ? 'enabled' : 'disabled';
     }
 }
