+++ src/Command/Configure/SessionHandler.php
@@ -38,6 +38,7 @@ use RuntimeException;
 class SessionHandler implements Module, ModuleUsage
 {
     use ModuleTrait;
+    use ConfigureHelperTrait;
 
     protected \Horde_Cli $cli;
     private ConfigManager $configManager;
@@ -482,42 +483,5 @@ class SessionHandler implements Module, ModuleUsage
         }
 
         return implode(', ', $servers) ?: 'localhost:11211';
-    }
-
-    /**
-     * Prompt for boolean value
-     *
-     * @param string $prompt Prompt text
-     * @param bool $default Default value
-     * @return bool User response
-     */
-    private function promptBoolean(string $prompt, bool $default): bool
-    {
-        $defaultStr = $default ? 'Y/n' : 'y/N';
-        $response = $this->cli->prompt("{$prompt} [{$defaultStr}]:", $default ? 'y' : 'n');
-        return strtolower($response) === 'y';
-    }
-
-    /**
-     * Parse boolean from string
-     *
-     * @param string $value String value
-     * @return bool Boolean value
-     */
-    private function parseBoolean(string $value): bool
-    {
-        $lower = strtolower($value);
-        return in_array($lower, ['true', '1', 'yes', 'y', 'on']);
-    }
-
-    /**
-     * Format boolean for display
-     *
-     * @param bool $value Boolean value
-     * @return string Formatted string
-     */
-    private function formatBoolean(bool $value): string
-    {
-        return $value ? 'yes' : 'no';
     }
 }
